backend/requirements.txt
Flask
flask-mysqldb
flask-mail
flask-cors
bcrypt
python-dotenv